Upgrade rkyv to 0.8 and remove bytecheck
The rkyv 0.8 version is a major update that brings many improvements and new
features. Therefore, I am eager for mmap-async to support rkyv 0.8. This patch
adapts the API to version 0.8, removing the previous 0.7 API.

Additionally, the key-value type in the test cases has been changed from
HashMap to BTreeMap, as the data bytes of HashMap are no longer fixed in
version 0.8. I suspect this is due to the unordered sorting of HashMap.

After switching to BTreeMap, everything works fine.